pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Ogaday Willers Moore <w.ogaday@gmail.com>"]
description = "Testing Workshop"
name = "testing-workshop"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.10"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
mkdocs-material = "^9.5.28"
pandas = "^2.2.2"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
vcrpy = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.1"
toml-sort = "^0.23.1"
ipython = "^8.26.0"

[tool.pytest.ini_options]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "serial"
]

[tool.tomlsort]
ignore_case = true
sort_table_keys = true

[tool.tomlsort.overrides."tool.poetry.dependencies"]
first = ["python"]
